pyproject.toml
[tool.poetry]
name = "casavant.jeff.perryfier"
version = "0.2.0"
description = "Maubot plugin for adding a brown fedora to existing pictures"
authors = ["Jeff Casavant <jeff.casavant@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8.0,<3.11"
maubot = "^0.3.1"
mautrix = "^0.15.8"
Pillow = "^9.0.1"
opencv-python = "^4.6.0"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black = "^21.8b0"
safety = "^1.10.3"
pylint = "^2.10.2"
isort = "^5.9.3"
toml = "^0.10.2"
"ruamel.yaml" = "^0.17.16"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint."MESSAGES_CONTROL"]
disable = [
    "bad-continuation",
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "too-few-public-methods",
    "too-many-locals",
    "too-many-statements"
]

[tool.pylint."FORMAT"]
max-line-length = 120

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.maubot]
maubot = "0.1.0"
modules = ["plugin"]
main_class = "PerryfierPlugin"
database = false
extra_files = [
  "res/img/perryhat.png",
]
